move tests for mkldnn conflict
